test_project/requirements.txt
pytest==6.2.5
pytest-selenium==3.0.0
